Set different field type when creating text field
